Prevent duplicate Check-In/Out and Web Clock-In/Out requests; correct Half/Full Day thresholds
The disabled-attribute-only protection wasn't sufficient: handlers had no
internal guard, so a second invocation (rapid click, or a click landing before
React committed the disabled re-render) still fired a real duplicate request,
surfacing as "You have already checked in today" / "No web clock-in found for
today". Added a synchronous useRef re-entrancy lock, checked/set as the first
line of every handler, across all entry points: AttendanceHeroBanner's
handlePunch/WebClockInRow (Check-In/Out, Web Clock-In/Out), the shared
WebClockInRequestModal, and AttendancePage's punch()/WebCheckInAction. Verified
live: invoking each handler 2-5 times back-to-back now produces exactly one
network request. Backend already has a DB unique constraint on
(employee_user_id, work_date) as a race-condition backstop; unchanged.

Also corrected the Half Day / Full Day thresholds: AttendanceProperties/
AttendanceConfigResponse's halfDayMaxHours/fullDayMinHours were typed as
whole-hour int (4/8) and couldn't represent the actual business rule of
3h30m/7h30m. Widened to double (3.5/7.5). Verified status is never finalized
while a session is still open (HALF_DAY is only ever set at actual checkout) —
that part was already correct.

+++ backend/src/main/java/com/nforce/onehr/dto/attendance/AttendanceConfigResponse.java
@@ -21,8 +21,9 @@ public class AttendanceConfigResponse {
     /** Null only if the caller has no Shift assigned — every employee is seeded with one (V95), so this is normally always set. */
     private LocalTime shiftEnd;
     private int lateGraceMinutes;
-    private int halfDayMaxHours;
-    private int fullDayMinHours;
+    // Fractional hours (3.5 = 3h30m) — mirrors AttendanceProperties.halfDayMaxHours/fullDayMinHours.
+    private double halfDayMaxHours;
+    private double fullDayMinHours;
     private int dailyBreakBudgetMinutes;
     /** java.time.DayOfWeek names, e.g. ["SATURDAY", "SUNDAY"]. */
     private List<String> weeklyOffDays;
